ccid: Fix overflow in Fi value
The Fi value can easily exceed 8bits. Even its default of 372 already
exceeds 256...

+++ ccid/ccid_device.h
@@ -19,8 +19,8 @@ struct msgb;
 
 struct ccid_pars_decoded {
 	/* global for T0/T1 */
-	uint8_t fi;
-	uint8_t di;
+	uint32_t fi;
+	uint32_t di;
 	enum ccid_clock_stop clock_stop;
 	bool inverse_convention;
